Routing for multiple API endpoints in a single Spring lambda
