+++ backend/main/src/main/java/sentiment/Response.java
@@ -1,23 +1,13 @@
 package sentiment;
 
+import com.fasterxml.jackson.annotation.JsonAnyGetter;
+
+import java.util.Map;
+
 /**
- * Response format for sentiment statistics.
+ * Response format for sentiment API.
  */
-public class Response {
-  /**
-   * The array of stats to return.
-   */
-  private OutgoingStat<?>[] stats;
-
-  public Response(OutgoingStat<?>[] stats) {
-    this.stats = stats;
-  }
-
-  public OutgoingStat<?>[] getStats() {
-    return this.stats;
-  }
-
-  public void setStats(OutgoingStat<?>[] stats) {
-    this.stats = stats;
-  }
+public abstract class Response {
+  @JsonAnyGetter
+  public abstract Map<String, Object> getData();
 }
