+++ backend/main/src/main/java/sentiment/stats/IncomingStat.java
@@ -1,4 +1,4 @@
-package sentiment;
+package sentiment.stats;
 
 import com.fasterxml.jackson.annotation.JsonAnySetter;
