+++ backend/main/src/test/java/sentiment/stats/StatsTests.java
@@ -1,4 +1,4 @@
-package sentiment;
+package sentiment.stats;
 
 import org.junit.Test;
 
@@ -20,7 +20,7 @@ public class StatsTests {
 	@Test
 	public void testCalculateStats() {
 
-    FunctionApplication functionApplication = new FunctionApplication();
+    StatsRequest request = new StatsRequest("", "", "2001-01-01T00:00:00.000Z", "2001-01-01T00:00:00.000Z", new IncomingStat[]{});
 
     List<Map<String, AttributeValue>> reviews = new ArrayList<Map<String, AttributeValue>>();
     reviews.add(populateReview("test store*test id", "2018-05-21T00:00:00.000Z", "1.0.0", "abcdef"));
@@ -30,7 +30,7 @@ public class StatsTests {
     rawReviews.setData("rawReviews", null);
     IncomingStat[] stats = new IncomingStat[] { rawReviews };
 
-    OutgoingStat<?>[] result = functionApplication.calculateStats(reviews, stats);
+    OutgoingStat<?>[] result = request.calculateStats(reviews, stats);
 
     String[] reviewsJson = new String[2];
     reviewsJson[0] = "{\"date\":\"2018-05-21T00:00:00.000Z\",\"reviewHash\":\"abcdef\",\"version\":\"1.0.0\",\"appIdStore\":\"test store*test id\"}";
@@ -57,14 +57,13 @@ public class StatsTests {
 
 	@Test
 	public void testProcessRawReviews() throws Exception {
-    FunctionApplication functionApplication = new FunctionApplication();
-
+    StatsRequest request = new StatsRequest("", "", "2001-01-01T00:00:00.000Z", "2001-01-01T00:00:00.000Z", new IncomingStat[]{});
 
     List<Map<String, AttributeValue>> reviews = new ArrayList<Map<String, AttributeValue>>();
     reviews.add(populateReview("test store*test id", "2018-05-21T00:00:00.000Z", "1.0.0", "abcdef"));
     reviews.add(populateReview("test store2*test id2", "2001-04-04T00:00:00.000Z", "0.0.1", "a1b2c3"));
 
-    OutgoingStat<String> result = functionApplication.processRawReviews(reviews);
+    OutgoingStat<String> result = request.processRawReviews(reviews);
 
     String[] reviewsJson = new String[2];
     reviewsJson[0] = "{\"date\":\"2018-05-21T00:00:00.000Z\",\"reviewHash\":\"abcdef\",\"version\":\"1.0.0\",\"appIdStore\":\"test store*test id\"}";
